refactor(admin): extract excel cell coercion to satisfy max-lines guard
excel-like-grid.tsx crossed 500 lines with the paste/coercion code;
coerceExcelValue moves to its own module (shared by commit + paste).

Refs #2402
